Renaming component back
